refactored maps.py to segregate gates & expressions related functions and constants
